test_jenkins/requirements.txt
python-jenkins
jupyter
pytest
